Fixes deserialization when a Content Type header is missing from the resposne
In the special case when a server does not return a content type header we can try to infer if the result was Json or XML from the request's Accept header. There doesn't seem to be a cleaner way to implement this as InterceptResponse occurs too late after the deserialization routine has already failed.